Set explicit plugin versions
The `workflow-aggregator:latest` plugin was updated to require Jenkins version 3.332.4, which caused the labs instance to fail to install a variety of different plugins. This caused various steps of the labs to break for numerous reasons.

Setting explicit plugin versions will ensure the labs don't randomly break. It will instead mean that periodically we'll want to come in and update plugin/jenkins versions